# Conflicts:
#	README.md
#	ph-bc/pom.xml
#	ph-charset/pom.xml
#	ph-cli/pom.xml
#	ph-collection/pom.xml
#	ph-commons/pom.xml
#	ph-commons/src/main/java/com/helger/commons/collection/impl/ICommonsIterable.java
#	ph-commons/src/main/java/com/helger/commons/random/RandomHelper.java
#	ph-commons/src/main/java/com/helger/commons/traits/IGenericAdderTrait.java
#	ph-config/pom.xml
#	ph-dao/pom.xml
#	ph-datetime/pom.xml
#	ph-graph/pom.xml
#	ph-jaxb/pom.xml
#	ph-json/pom.xml
#	ph-json/src/main/java/com/helger/json/serialize/JsonReader.java
#	ph-less-commons/pom.xml
#	ph-matrix/pom.xml
#	ph-scopes/pom.xml
#	ph-security/pom.xml
#	ph-settings/pom.xml
#	ph-tree/pom.xml
#	ph-wsclient/pom.xml
#	ph-xml/pom.xml
#	ph-xml/src/main/java/com/helger/xml/microdom/serialize/MicroSerializer.java
#	pom.xml

+++ ph-commons/src/main/java/com/helger/commons/io/stream/NonBlockingBufferedInputStream.java
@@ -372,7 +372,7 @@ public class NonBlockingBufferedInputStream extends WrappedInputStream
     }
 
     final long nSkipped = nAvail < nBytesToSkip ? nAvail : nBytesToSkip;
-    m_nPos += nSkipped;
+    m_nPos += (int) nSkipped;
     return nSkipped;
   }
